create publisher in separate process
To send data across the process boundary, a publisher/subscriber pair
needs to be created together with stub/skeleton wrappers for each
side.  This commit begins the work.

Current skeleton code creates a thread to service the socket, which is
a natural decision to make.  Unfortunately, QML (and other UI kits)
have limit UI interaction to the gui thread.  When the skeleton thread
attempts to post metric descriptions to a GUI element, an assertion is
thrown.

This needs to be sorted out in one of a couple ways:

 1) all input processing happens on the main thread.
    - this implies that a set of sockets needs to be checked with
      select() on a regular basis by the main thread.  Or, a thread
      blocks on the fd set, and posts an event to the main thread when
      any socket has data.

    - a limitation of this approach has to do with synchronous calls.
      remote synchronous calls are not really possible in this
      paradigm, because they would have to be limited to the gui
      thread, and would block the gui.

 2) all GUI events are posted to the GUI thread via a slot/signal.
    This allows more conventional RPC, but will result in a tedious
    layer between the GUI and the rest of the app.  This seems similar
    to the concept of a QObject View-model.