fix UGLY proliferation of slot/signals around stub/skeleton implementations.
The rule that has to be followed is that you can't post an attribute
or send a signal to the GUI thread from anything but the main thread.
To get around this, I was trying to have the skeleton threads emit
signals that the subscriber/producer models could catch on the main
thread, where they were created.

It is simpler to just call through to those models off-thread, but in
a thread-safe way.  The implementation of the virtual methods
satisfied by the model can then translate to QObjects in the UI
thread.

This is accomplished by creating a slot/signal pair within the model
object, and connecting them when the object is constructed.
Model notification (publication etc) can be received in any thread,
but the model has to emit a signal to itself so that subsequent QML
attributes can be updated in the main thread.  The signal is handled
on the UI thread as per standard Qt.

This solution collapses the model-view (QObjects) and the
model (Grafips subscribers etc) into a single class.