Remove QObject from publisher/provider classes
publishers and providers will be instantiated within the
fips-instrumented application.  Since fips is not going to link
against Qt, we can have Qt in those classes.

The Qt types were necessary to get data flowing within a single
application.  With this change, the stubs became the QObjects, since
they are now instantiated in QML.