drop unnecessary GetDescriptions publisher method
To introduce a Metrics Router, we need to make the list of displayed
metrics dependent on a list of filters that a graph will be
instantiated with.  The graph no longer retrieves its metrics
synchronously from the publisher (which was substandard, anyways).

This change makes the metrics list reactive, consisting solely of the
metrics that are passed asynchronously upon subscribe.