finalize drag/drop functionality
It turns out that the clipping of the drag item has everything to do
with the parenting of the item.  Re-parenting to a larger scope allows
the item to be drug anywhere, but requires recalculation of
coordinates in terms of the wider scope.

Altering coordinates within a drag event did not reposition the item
succesfully, so it was broken up:  onPressed sets the location, and
the drag event subsequently makes the item visible.