Request the connected address from the server socket.
When subscribing, the client has to provide it's ip address to the
publisher.  I was using localhost, or hostname.local, but these failed
if the system was not remote or if there was no connected network
interface.

The server socket knows the ip of the connected client, because it
accepts the connection.  The subscribe protocol was enhanced so the
client can make a synchronous GetClientAddress from the server, and
use the response in the subsequent subscribe message.

There are a couple of ways to skin this cat, especially since we have
only one subscriber at the moment.

The same change needs to be made for the control subscription.