improve the handling of ip in subscribe messages
Previously, the address of the subscriber was sent in the subscribe
message.  This was problematic, because the subscriber did not know
which network interface it had used to connect to the publisher.
Using localhost broke two-system setups (the expected use case), and
using hostname.local broke if none of the network interfaces were
enabled (standalone / unit test use case).

Sending the address of the subscriber in the subscribe message turns
out to be unnecessary, because the publisher already knows the network
address of the subscriber.  It can obtain that information on
accept().  Grafips::Socket was refactored to retain and expose that
information, and expose it to callers.